fix entity extraction
